save place renders to a renders folder and clean up some more code
