+++ models/export_model.py
@@ -27,7 +27,7 @@ def export_model():
     full_filename = parts[-1]
     full_filename_parts = full_filename.split(".")
     filename_name = full_filename_parts[-2]
-    this_place_name = filename_name
+    this_model_name = filename_name
 
     # Get model names from folder names (folders with index files)
     model_names = []
@@ -95,12 +95,12 @@ def export_model():
         export_animations=True,
         export_lights=False,
         use_selection=True,
-        filepath=parent_folder_path + os.sep + this_place_name + ".glb",
+        filepath=parent_folder_path + os.sep + this_model_name + ".glb",
     )
 
     # Save index file
-    with open(parent_folder_path + os.sep + this_place_name + ".ts", "w") as file:
-        file.write(f'import modelFile from "./{this_place_name}.glb";\n\n')
+    with open(parent_folder_path + os.sep + this_model_name + ".ts", "w") as file:
+        file.write(f'import modelFile from "./{this_model_name}.glb";\n\n')
 
         # animation_names
         # bone_names
@@ -157,11 +157,11 @@ def export_model():
 
         # grandparent_folder_path
 
-    # Save all places index file
-    with open(grandparent_folder_path + os.sep + "places.ts", "w") as file:
+    # Save all models index file (models.ts)
+    with open(grandparent_folder_path + os.sep + "models.ts", "w") as file:
         for looped_name in model_names:
             file.write(
-                f'import {{ modelInfo as {looped_name}Info }} from "./{looped_name}";\n'
+                f'import {{ modelInfo as {looped_name}Info }} from "./{looped_name}/{looped_name}";\n'
             )
 
         file.write("\n")
