+++ places/combine_videos.py
@@ -1,10 +1,13 @@
 import os
+import shutil
 import subprocess
 
 
-def combine_videos(parent_folder_path, camera_names, segments_for_cams):
+def combine_videos(
+    renders_folder_path, place_folder_path, camera_names, segments_for_cams
+):
     # join_color_vids.txt
-    with open(parent_folder_path + os.sep + "join_color_vids.txt", "w") as file:
+    with open(renders_folder_path + os.sep + "join_color_vids.txt", "w") as file:
         # Import videos
         file.write(
             "# tells the video joiner the names and order of the color videos \n"
@@ -15,7 +18,7 @@ def combine_videos(parent_folder_path, camera_names, segments_for_cams):
         file.write("\n")
 
     # join_depth_vids.txt
-    with open(parent_folder_path + os.sep + "join_depth_vids.txt", "w") as file:
+    with open(renders_folder_path + os.sep + "join_depth_vids.txt", "w") as file:
         # Import videos
         file.write(
             "# tells the video joiner the names and order of the depth videos \n"
@@ -25,17 +28,26 @@ def combine_videos(parent_folder_path, camera_names, segments_for_cams):
                 file.write(f"file '{looped_cam_name}_{segment_name}_depth.mp4'\n")
         file.write("\n")
 
-    combineVideosCommand_color = f'C:\\ffmpeg -f concat -i "join_color_vids.txt" -c copy -y "{parent_folder_path}{os.sep}color.mp4"  -hide_banner -loglevel error'
-    combineVideosCommand_depth = f'C:\\ffmpeg -f concat -i "join_depth_vids.txt" -c copy -y "{parent_folder_path}{os.sep}depth.mp4"  -hide_banner -loglevel error'
+    combineVideosCommand_color = f'C:\\ffmpeg -f concat -i "join_color_vids.txt" -c copy -y "{renders_folder_path}{os.sep}color.mp4"  -hide_banner -loglevel error'
+    combineVideosCommand_depth = f'C:\\ffmpeg -f concat -i "join_depth_vids.txt" -c copy -y "{renders_folder_path}{os.sep}depth.mp4"  -hide_banner -loglevel error'
 
-    video_quality = "25"
+    video_quality = "23"
     keyframes = "1"
 
-    combineColorAndDepthVertically = f'C:\\ffmpeg -i "{parent_folder_path}{os.sep}color.mp4" -i "{parent_folder_path}{os.sep}depth.mp4" -filter_complex vstack=inputs=2 -vcodec libx264 -crf {video_quality} -g {keyframes} -y -movflags faststart "{parent_folder_path}{os.sep}backdrops.mp4" -hide_banner -loglevel error'
+    combineColorAndDepthVertically = f'C:\\ffmpeg -i "{renders_folder_path}{os.sep}color.mp4" -i "{renders_folder_path}{os.sep}depth.mp4" -filter_complex vstack=inputs=2 -vcodec libx264 -crf {video_quality} -g {keyframes} -y -movflags faststart "{renders_folder_path}{os.sep}backdrops.mp4" -hide_banner -loglevel error'
 
     # subprocess.run(f"cd {parent_folder_path}")
-    subprocess.run(combineVideosCommand_color, cwd=parent_folder_path)
-    subprocess.run(combineVideosCommand_depth, cwd=parent_folder_path)
-    subprocess.run(combineColorAndDepthVertically, cwd=parent_folder_path)
+    subprocess.run(combineVideosCommand_color, cwd=renders_folder_path)
+    subprocess.run(combineVideosCommand_depth, cwd=renders_folder_path)
+    subprocess.run(combineColorAndDepthVertically, cwd=renders_folder_path)
 
-    # TODO Delete the text files here? (are the subprocess done?)
+    # Delete the text files here? (are the subprocess done?)
+    os.remove(renders_folder_path + os.sep + "join_color_vids.txt")
+    os.remove(renders_folder_path + os.sep + "join_depth_vids.txt")
+
+    # move the backdrops.mp4 to the place folder, if it exists
+    if os.path.exists(renders_folder_path + os.sep + "backdrops.mp4"):
+        shutil.move(
+            renders_folder_path + os.sep + "backdrops.mp4",
+            place_folder_path + os.sep + "backdrops.mp4",
+        )
