+++ places/custom_render_video.py
@@ -9,7 +9,7 @@ import shutil
 def custom_render_video(
     camName,
     segmentName,
-    parent_folder_path,
+    renders_folder_path,
     segments_info,
     chosen_framerate,
     fileNamePost="",
@@ -17,10 +17,10 @@ def custom_render_video(
     scene = get_scene()
 
     # segments_info, chosen_framerate paramerter
-    # get_parent_folder_path()
+    # get_renders_folder_path()
 
     frame_image_folder_path = (
-        f"{parent_folder_path}{os.sep}{camName}_{segmentName}{fileNamePost}"
+        f"{renders_folder_path}{os.sep}{camName}_{segmentName}{fileNamePost}"
     )
 
     segment_info = segments_info[segmentName]
@@ -28,9 +28,9 @@ def custom_render_video(
     amount_of_frames_for_segment = segment_info.frameEnd - segment_info.frameStart
     amount_of_frames_saved = int(round(amount_of_frames_for_segment / scene.frame_step))
     video_output_path = (
-        f"{parent_folder_path}{os.sep}{camName}_{segmentName}{fileNamePost}"
+        f"{renders_folder_path}{os.sep}{camName}_{segmentName}{fileNamePost}"
     )
-    video_quality = "25"
+    video_quality = "23"
     keyframes = "1"
 
     # renders each image frame manually so they can have sequential names like image0000 image0001 instead of frame0000, frame0006
@@ -49,7 +49,7 @@ def custom_render_video(
         f"{frame_image_path}{str(amount_of_frames_saved ).zfill(4)}.png",
     )
 
-    makeVideoCommand = f'C:\\ffmpeg -framerate {chosen_framerate} -f image2 -i "{frame_image_path}%04d.png" -vcodec libx264 -crf {video_quality} -g {keyframes} -vf "fps={chosen_framerate},format=yuv420p,scale=1280:720" -y -movflags faststart "{video_output_path}.mp4" -hide_banner -loglevel error'
+    makeVideoCommand = f'C:\\ffmpeg -framerate {chosen_framerate} -f image2 -i "{frame_image_path}%04d.png" -vcodec libx264 -crf {video_quality} -g {keyframes} -vf "fps={chosen_framerate},format=yuv420p,scale=1600:900" -y -movflags faststart "{video_output_path}.mp4" -hide_banner -loglevel error'
     # -movflags faststart is for starting video fast on web?
     # NOTE -g might be depricated soon?
