+++ places/places.py
@@ -1,19 +1,16 @@
-from math import radians
-
 import bpy
 
-# custom ui imports
-from mathutils import Euler, Vector
-
-from .place_info import PlaceInfo
-from .update_items_and_variables import update_items_and_variables
-
 from ..get_things import get_collections, get_scene, get_view_layer
+from .cam_background import setup_cam_background
 from .collections import (
     add_collection_to_exportable_collection,
     add_collection_to_scene,
     enable_all_child_collections,
 )
+from .place_info import PlaceInfo
+from .probes import setup_camera_probes
+from .update_items_and_variables import update_items_and_variables
+
 
 # ----------------------------------------------------------------------------------------
 # Setup everything
@@ -38,176 +35,6 @@ def setup_video_rendering(place_info: PlaceInfo):
 
     # for faster performance
     scene.render.use_persistent_data = True
-
-
-def toggle_world_volume(isToggled=True):
-    global original_world_volume_color
-
-    world_nodes = bpy.data.worlds["World"].node_tree.nodes
-
-    volume_inputs = []
-
-    if "Principled Volume" in world_nodes:
-        volume_inputs = world_nodes["Principled Volume"].inputs
-    elif "Volume Scatter" in world_nodes:
-        volume_inputs = world_nodes["Volume Scatter"].inputs
-    else:
-        return
-
-    density_input = volume_inputs[2]
-    volume_color_input = volume_inputs[0]
-
-    if isToggled:
-        if not original_world_volume_color:
-            original_world_volume_color = volume_color_input.default_value
-        density_input.default_value = 0.0013
-        volume_color_input.default_value = original_world_volume_color
-    else:
-        density_input.default_value = 0
-        original_world_volume_color = volume_color_input.default_value
-        volume_color_input.default_value = (0, 0, 0, 1)
-
-
-def toggle_depth_hidden_objects(isToggled=True):
-    view_layer = get_view_layer()
-    # NOTE HAVE to loop through view_layer ?? , can't loop through collection.children ohwell
-    for collection in view_layer.layer_collection.children["Details"].children:
-        if collection.name == "hidden_to_depth":
-            collection.exclude = not isToggled
-
-
-def toggle_depth_visible_objects(isToggled=True):
-    view_layer = get_view_layer()
-    # NOTE HAVE to loop through view_layer ?? , can't loop through collection.children ohwell
-    for collection in view_layer.layer_collection.children["Details"].children:
-        if collection.name == "visible_to_depth":
-            collection.exclude = not isToggled
-
-
-def toggle_probe_visible_objects(isToggled=True):
-    view_layer = get_view_layer()
-    # NOTE HAVE to loop through view_layer ?? , can't loop through collection.children ohwell
-    for collection in view_layer.layer_collection.children["Details"].children:
-        if collection.name == "visible_to_probe":
-            collection.exclude = not isToggled
-
-
-# not needed! might help if stopping "make videos" before it finshed
-def reenable_all_meshes(place_info: PlaceInfo):
-    global meshnames_that_were_disabled_in_render
-
-    for cam_name in place_info.hidden_meshes_for_cams:
-        for mesh_name in place_info.hidden_meshes_for_cams[cam_name]:
-            mesh_object = bpy.data.objects[mesh_name]
-            mesh_object.hide_render = False
-            mesh_object.visible_camera = True
-
-
-def reenable_hidden_meshes():
-    global meshnames_that_were_disabled_in_render
-    # re-enable previously hidden meshes
-    for mesh_name in meshnames_that_were_disabled_in_render:
-        mesh_object = bpy.data.objects[mesh_name]
-        mesh_object.visible_camera = True
-        mesh_object.hide_render = False
-    meshnames_that_were_disabled_in_render = []
-
-
-def hide_meshes_for_camera(place_info: PlaceInfo, cam_name, isDepth=False):
-    global meshnames_that_were_disabled_in_render
-    # hide meshes to camera that should be hidden
-    if cam_name in place_info.hidden_meshes_for_cams:
-        for mesh_name in place_info.hidden_meshes_for_cams[cam_name]:
-            print(mesh_name)
-            mesh_object = bpy.data.objects[mesh_name]
-            if isDepth:
-                mesh_object.hide_render = True
-            else:
-                mesh_object.visible_camera = False
-            meshnames_that_were_disabled_in_render.append(mesh_name)
-    else:
-        print(f"{cam_name} has no hidden meshes :)")
-
-
-def setup_probe_rendering():
-    scene = get_scene()
-
-    scene.render.resolution_x = 256
-    scene.render.resolution_y = 128
-
-    scene.render.image_settings.file_format = "HDR"
-    scene.node_tree.nodes["Denoise"].use_hdr = True
-
-
-def setup_camera_probes():
-    collections = get_collections()
-    # Loop through current cameras and add probes if needed
-
-    for cam_collection in collections["cameras"].children:
-        probe_object = None
-        camera_object = None
-        for child_object in cam_collection.objects:
-            if child_object.type == "CAMERA":
-                if child_object.data.type == "PANO":
-                    probe_object = child_object
-                else:
-                    child_object.name = cam_collection.name
-                    camera_object = child_object
-        if probe_object is None:
-            bpy.ops.object.camera_add(
-                location=[0, 10, 20],
-                rotation=[radians(90), radians(0), radians(180)],
-            )
-            # https://blender.stackexchange.com/questions/132112/whats-the-blender-2-8-command-for-adding-an-object-to-a-collection-using-python
-            # our created camera is the active one
-            probe_object = bpy.context.active_object
-            # Remove (active) object from all collections not used in a scene
-            bpy.ops.collection.objects_remove_all()
-            # add it to our specific collection
-            cam_collection.objects.link(probe_object)
-            # https://blender.stackexchange.com/questions/26852/python-move-object-on-local-axis
-            # 15 blender units in z-direction
-            dist_z = Vector((0.0, 0.0, -15))
-            rotationMAT = camera_object.rotation_euler.to_matrix()
-            rotationMAT.invert()
-            # project the vector to the world using the rotation matrix
-            zVector = dist_z @ rotationMAT
-            probe_object.location = camera_object.location + zVector
-            # probe_object.location = camera_object.location + Vector((0, -3, -2))
-
-        probe_object.rotation_euler = Euler(
-            (radians(90), radians(0), radians(180)), "XYZ"
-        )
-        probe_object.data.type = "PANO"
-        probe_object.data.cycles.panorama_type = "EQUIRECTANGULAR"
-        probe_object.name = f"{camera_object.name}_probe"
-        # show an axis instead of the camera
-        probe_object.show_axis = True
-        probe_object.data.display_size = 0.01
-
-        # add custom properties to the camera
-        # camera_object.
-
-
-def setup_cam_background():
-    scene = get_scene()
-
-    try:
-        # get background image struct
-        active_cam = scene.camera.name
-        bg_images = bpy.data.objects[active_cam].data.background_images.items()
-        # get background image data, if it exists in struct
-        found_background_image = bg_images[0][1].image
-        scene.node_tree.nodes["image_node"].image = found_background_image
-        scene.node_tree.nodes["switch_background"].check = True
-        scene.render.film_transparent = True
-        scene.view_settings.view_transform = "Standard"
-
-        # image_scale = bg_images[0][1].scale
-    except:
-        scene.node_tree.nodes["switch_background"].check = False
-        scene.render.film_transparent = False
-        print("No Background Found")
 
 
 def setup_place(place_info: PlaceInfo, the_render_quality, the_framerate):
